app/requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
dill==0.3.7
distro==1.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
ffmpeg==1.4
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gotrue==1.3.0
greenlet==3.0.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.19.4
idna==3.4
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai==1.3.5
packaging==23.2
pandas==2.1.3
passlib==1.7.4
pip-autoremove==0.10.0
postgrest==0.13.0
psutil==5.9.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
realtime==1.0.0
regex==2023.10.3
requests==2.31.0
rsa==4.9
safetensors==0.4.0
semantic-version==2.10.0
setuptools-rust==1.8.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.1.0
supafunc==0.3.1
sympy==1.12
tiktoken==0.5.1
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.1
tqdm==4.66.1
trash-cli==0.23.11.10
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
websockets==10.4
xxhash==3.4.1
yarl==1.9.3
